Organize speech samples into grpc vs REST
