+++ speech/README.md
@@ -1,0 +1,14 @@
+# Google Cloud Speech API Samples
+
+These samples show how to use the [Google Cloud Speech API][speech-api]
+to transcribe audio files, as well as live audio from your computer's
+microphone.
+
+* [`api-client/`](api-client) contains samples that use the [Google API Client
+  Library for Python][rest-client] to make REST calls to the API.
+* [`grpc/`](grpc) contains samples that use the more-efficient (though sometimes
+  more complex) [GRPC][grpc] API. The GRPC API also allows streaming requests.
+
+[speech-api]: http://cloud.google.com/speech
+[rest-client]: https://developers.google.com/api-client-library/python/
+[grpc]: http://grpc.io
