+++ speech/api-client/resources/audio.raw
