+++ speech/api-client/transcribe.py
