+++ speech/grpc/resources/quit.raw
