+++ speech/grpc/transcribe.py
